fix: 무한스크롤 렌더링하는 위치 수정
